The comments were taken into account

+++ src/main/resources/db.migration/query.sql
@@ -9,14 +9,14 @@ create table users (
 );
 
 create table roles (
-                       id                    integer auto_increment,
+                       id                    bigint auto_increment,
                        name                  varchar(50) not null,
                        primary key (id)
 );
 
 CREATE TABLE users_roles (
                              user_id               bigint not null,
-                             role_id               int not null,
+                             role_id               bigint not null,
                              primary key (user_id, role_id),
                              foreign key (user_id) references users (id),
                              foreign key (role_id) references roles (id)
